quickDecoder/requirements.txt
pyperclip==1.7.0
python-baseconv==1.2.1
